refactor: move PlanBasedEngine to delta_kernel_default_engine
Relocate the plan-based engine handlers and PlanBasedEngine from the
kernel crate into default-engine behind the declarative-plans feature,
updating FFI imports accordingly. No behavior changes.
