+++ kernel/src/engine/mod.rs
@@ -43,11 +43,6 @@ pub(crate) mod arrow_utils;
 #[cfg(all(feature = "internal-api", feature = "arrow-expression"))]
 pub use self::arrow_utils::{parse_json, to_json_bytes};
 
-// The plan executor support modules read Arrow data (`arrow_utils`, `arrow_data`), so they
-// require the Arrow engine base in addition to the declarative-plans IR.
-#[cfg(all(feature = "declarative-plans", feature = "default-engine-base"))]
-pub mod plans;
-
 #[cfg(test)]
 pub(crate) mod sync;
